Speed up iterations by x8 times faster, fix minor bug in data saving
New data

+++ main/statistics/main.py
@@ -11,15 +11,15 @@ def solutionStatistics(solution):
     placed = pd.DataFrame(solution["placed"])
     discard = pd.DataFrame(solution["discard"]) if len(solution["discard"]) else placed.iloc[0:1].apply(lambda x: 0, axis=0)
     return {"iteration": solution["iteration"], "time": solution["time"],
-            "p_mean_weight": float(placed.weight.mean()), "d_mean_weight": discard.weight.mean(),
-            "p_total_weight": float(placed.weight.sum()), "d_total_weight": discard.weight.sum(),
-            "p_mean_taxability": float(placed.weight.mean()), "d_mean_taxability": discard.weight.mean(),
-            "p_total_taxability": float(placed.taxability.sum()), "d_total_taxability": discard.taxability.sum(),
-            "p_mean_volume": float(placed.volume.mean()), "d_mean_volume": discard.volume.mean(),
-            "p_cumulative_priority": int(placed.priority.sum()), "p_mean_priority": placed.priority.mean(),
+            "p_mean_weight": float(placed.weight.mean()), "d_mean_weight": float(discard.weight.mean()),
+            "p_total_weight": float(placed.weight.sum()), "d_total_weight": float(discard.weight.sum()),
+            "p_mean_taxability": float(placed.weight.mean()), "d_mean_taxability": float(discard.weight.mean()),
+            "p_total_taxability": float(placed.taxability.sum()), "d_total_taxability": float(discard.taxability.sum()),
+            "p_mean_volume": float(placed.volume.mean()), "d_mean_volume": float(discard.volume.mean()),
+            "p_cumulative_priority": int(placed.priority.sum()), "p_mean_priority": int(placed.priority.mean()),
             "d_max_priority": int(discard.priority.max()),
-            "used_volume": placed.volume.sum() / solution["truck"]["volume"],
-            "used_weight": placed.weight.sum() / solution["truck"]["tonnage"]
+            "used_volume": float(placed.volume.sum() / solution["truck"]["volume"]),
+            "used_weight": float(placed.weight.sum() / solution["truck"]["tonnage"])
             }
